bms_lv_rust
Macro info

macro_rules! info {
    ($($arg:tt)*) => { ... };
}

Logs formatted strings over USB CDC-ACM.

This macro takes a format string and arguments similar to core::format!, stores the formatted string in a static buffer, and triggers the USB task to send it to the host.

Example

info!("Battery voltage: {}mV", voltage);

Notes

Messages are saved via a global static buffer SEND_BUF to be eventually sent by the write_usb task.