bms_lv_rust::usb_serial
Module serial_usb

USB Serial Communication module using the CDC-ACM class with Embassy and STM32.

This module provides asynchronous USB CDC functionality using the embassy-usb stack. It is primarily used to send debug or runtime information to a host PC over USB.

Structs

Disconnected

Represents a USB disconnection or endpoint error.

Irqs 🔒

UsbDevice

A USB CDC-ACM device instance with Embassy integration.

Statics

BOS_DESCRIPTOR 🔒

CONFIG_DESCRIPTOR 🔒

CONTROL_BUF 🔒

EP_OUT_BUF 🔒

MSOS_DESCRIPTOR 🔒

SEND_BUF

Static buffer used to pass outgoing messages from other parts of the system to the USB task.

STATE 🔒

Functions

OTG_FS 🔒 ⚠

usb_run_task

Runs the USB device task. Should be spawned once at startup.